+++ packages/react-ui/storybook/stories/widgetsUI/LegendWidgetUI.stories.js
@@ -41,6 +41,7 @@ const Template = ({ ...args }) => {
   );
 };
 
+
 const LegendTemplate = () => {
   const layers = [
     {
@@ -61,6 +62,22 @@ const LegendNotFoundTemplate = () => {
       id: 0,
       title: 'Single Layer',
       visible: true
+    }
+  ];
+  return <LegendWidgetUI layers={layers}></LegendWidgetUI>;
+};
+
+const LegendWithOpacityTemplate = () => {
+  const layers = [
+    {
+      id: 0,
+      title: 'Single Layer',
+      visible: true,
+      showOpacityControl: true,
+      opacity: 0.5,
+      legend: {
+        children: <div>Your Content</div>,
+      }
     }
   ];
   return <LegendWidgetUI layers={layers}></LegendWidgetUI>;
@@ -258,6 +275,7 @@ export const SingleLayer = LegendTemplate.bind({});
 export const MultiLayer = LegendMultiTemplate.bind({});
 export const MultiLayerCollapsed = LegendMultiTemplateCollapsed.bind({});
 export const NotFound = LegendNotFoundTemplate.bind({});
+export const LegendWithOpacityControl = LegendWithOpacityTemplate.bind({});
 export const Categories = LegendCategoriesTemplate.bind({});
 export const CategoriesAsStroke = LegendCategoriesStrokeTemplate.bind({});
 export const Icon = LegendIconTemplate.bind({});